Split aster into two crates
Unfortunately we cannot use features for this because
as of right now cargo doesn't support features that
conflict with each other, and so if one dependency
depends on a "nightly" feature, and another on a "syntex"
feature, cargo will enable both features on the project.

Another way to avoid this would be to have just one
"aster" crate, which has submodules "syntex" and "nightly",
but that feels a little ugly.